requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
bootstrap4==0.1.0
crispy-bootstrap4==2023.1
crispy-bootstrap5==2023.10
Django==4.2.9
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap3==23.6
django-bootstrap3-multidatepicker==0.1.5
django-bootstrap5==23.4
django-crispy-forms==2.1
django-datetime-widget==0.9.3
django-datetimepicker==3.14
mssql-django==1.4
numpy==1.26.3
pandas==2.2.0
pydantic==2.6.0
pydantic_core==2.16.1
pyodbc==5.0.1
python-dateutil==2.8.2
pytz==2023.4
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
whitenoise==6.6.0
